+++ src/app/store/routes/receipt/receipt-card/store-receipt-card.component.spec.ts
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2021 The Trébol eCommerce Project
+ *
+ * This software is released under the MIT License.
+ * https://opensource.org/licenses/MIT
+ */
+
+import { CommonModule } from '@angular/common';
+import { Component, Input } from '@angular/core';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { MatCardModule } from '@angular/material/card';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { RouterTestingModule } from '@angular/router/testing';
+import { EMPTY } from 'rxjs';
+import { API_SERVICE_INJECTION_TOKENS } from 'src/app/api/api-service-injection-tokens';
+import { IReceiptPublicApiService } from 'src/app/api/receipt-public-api.iservice';
+import { StoreReceiptCardComponent } from './store-receipt-card.component';
+
+@Component({ selector: 'app-centered-mat-spinner' })
+class MockCenteredMatSpinnerComponent { }
+
+@Component({ selector: 'app-store-receipt-details-table' })
+class MockReceiptDetailsTableComponent {
+  @Input() details: any[];
+}
+
+describe('StoreReceiptCardComponent', () => {
+  let component: StoreReceiptCardComponent;
+  let fixture: ComponentFixture<StoreReceiptCardComponent>;
+  let mockReceiptApiService: Partial<IReceiptPublicApiService>;
+
+  beforeEach(waitForAsync(() => {
+    mockReceiptApiService = {
+      fetchTransactionReceiptByToken() { return EMPTY; }
+    };
+
+    TestBed.configureTestingModule({
+      imports: [
+        CommonModule,
+        NoopAnimationsModule,
+        RouterTestingModule,
+        MatCardModule,
+        MatIconModule,
+        MatListModule,
+      ],
+      declarations: [
+        StoreReceiptCardComponent,
+        MockCenteredMatSpinnerComponent,
+        MockReceiptDetailsTableComponent
+      ],
+      providers: [
+        { provide: API_SERVICE_INJECTION_TOKENS.receipt, useValue: mockReceiptApiService }
+      ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(StoreReceiptCardComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});